requirements.txt
fastapi
uvicorn
pydantic
presidio-analyzer
presidio-anonymizer
spacy
streamlit
requests
python-docx
fastmcp
Pillow
mistralai
python-multipart
openpyxl
cachetools
azure-cosmos
python-dotenv
aiohttp
httpx
PyJWT[crypto]
mcp